% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedCV.R
\name{functionalDetectionStats}
\alias{functionalDetectionStats}
\title{Given a vector functional (true) attribute associations and a vector of positive associations,
returns detection statistics like true positive rate, recall and precision.}
\usage{
functionalDetectionStats(functional, positives)
}
\arguments{
\item{functional}{A vector functional (true) attributes}

\item{positives}{A vector of positive associations}
}
\value{
A list with:
\describe{
\item{TP}{True Positive}
\item{FP}{False Positive}
\item{FN}{False Negative}
\item{TPR}{True Positive Rate}
\item{FPR}{False Positive Rate}
\item{precision}{Fraction of relevant features among the retrieved features}
\item{recall}{Fraction of total relevant features that were actually retrieved}
}
num.samples <- 100
num.variables <- 100
pct.signals <- 0.1
label <- "class"
sim.data <- createSimulation(num.samples = num.samples,
num.variables = num.variables,
pct.signals = pct.signals,
sim.type = "mainEffect",
label = label,
verbose = FALSE)
rnCV.results <- regular_nestedCV(train.ds = sim.data$train,
validation.ds = sim.data$holdout,
label = label,
is.simulated = TRUE,
ncv_folds = c(10, 10),
param.tune = FALSE,
learning_method = "rf",
importance.algorithm = "ReliefFbestK",
num_tree = 500,
verbose = FALSE)
functional <- data.sets$signal.names
rncv.positives <- rncv_result$Features
rncv.detect <- functionalDetectionStats(functional, rncv.positives)
}
\description{
Given a vector functional (true) attribute associations and a vector of positive associations,
returns detection statistics like true positive rate, recall and precision.
}
\seealso{
Other nestedCV: 
\code{\link{consensus_nestedCV}()},
\code{\link{regular_nestedCV}()}
}
\concept{nestedCV}
